carried on with frontend
added support for a graph, need to add in more features, make it look better and then work on implementing real data into the app. The current method would be to have api calls for individual stock data and then sum it up, graphql would make it easier to batch this but i sstill would need to add up all the data, from seperate sql queries, idk how consuming that is but it would be a start.